🚚 Move: user 상수 파일명 수정 #20
